# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Medclara Backend API
  description: |
    # Medclara Clinical Documentation Platform API
    
    Complete API specification for the Medclara backend, including:
    - **Authentication** - User login and session management
    - **Patients** - Patient data management
    - **Templates** - Clinical note templates
    - **Sessions** - Scribe session management
    - **Recordings** - Audio recording upload and processing
    - **Chunked Upload** - Large file upload with resumable chunks
    - **Server-Sent Events (SSE)** - Real-time processing status updates
    - **Notes** - Clinical note creation and management
    - **Workspace** - Scribe workspace API
    
    ## Real-Time Processing with SSE
    
    The Medclara API uses Server-Sent Events (SSE) for real-time recording processing updates.
    This provides 99% fewer network requests and 20x faster response times compared to polling.
    
    ### Complete Recording Workflow
    1. **POST** `/api/recordings/chunked/init` - Initialize upload session
    2. **POST** `/api/recordings/chunked/upload` - Upload audio chunks (repeat)
    3. **POST** `/api/recordings/chunked/finalize` - Start processing
    4. **GET** `/api/recordings/{recordingId}/stream` - Stream processing events (SSE)
    5. Process events: `processing` → `completed` or `failed`
    
    ### Benefits of SSE Over Polling
    - **99% fewer requests** - One persistent connection vs. polling every 500ms
...
    
  version: '1.0.0'
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          examples:
            - unauthorized
            - not_found
            - invalid_request
            - server_error
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
    Session:
      type: object
      required:
        - sessionId
        - status
        - templateId
        - createdAt
      properties:
        sessionId:
          type: string
          format: uuid
          description: Unique session identifier
        status:
          type: string
          enum: [initialized, active, completed]
        templateId:
          type: string
          format: uuid
          description: Template used for this session
        patientId:
          type: string
          format: uuid
          description: Associated patient
        encounterId:
          type: string
          description: Associated encounter
        noteSections:
          type: object
          description: Current note content organized by section
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        transcriptSegments:
          type: array
          items:
            $ref: '#/components/schemas/TranscriptSegment'
    CreateSessionRequest:
      type: object
      required:
        - templateId
      properties:
        templateId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        encounterId:
          type: string
        sessionId:
          type: string
          description: Optional custom session ID
        metadata:
          type: object
          additionalProperties: true
    BindPatientRequest:
      type: object
      required:
        - patientId
      properties:
        patientId:
          type: string
          format: uuid
        encounterId:
          type: string
    UpdateNoteSectionRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Content for this note section
        encounterId:
          type: string

    # ========================================================================
    # Recording Schemas
    # ========================================================================
...
    TranscriptSegment:
      type: object
      required:
        - id
        - text
        - start_time_seconds
        - end_time_seconds
      properties:
        id:
          type: string
          format: uuid
        speaker:
          type: string
          enum: [provider, patient]
        text:
          type: string
          description: Transcribed text
        start_time_seconds:
          type: number
          description: Start time in recording
        end_time_seconds:
          type: number
          description: End time in recording
        confidence:
          type: number
          minimum: 0
          maximum: 1

    # ========================================================================
    # Chunked Upload Schemas
    # ========================================================================
...
paths:
...
  /api/sessions:
    post:
      tags:
        - Sessions
      summary: Create session
      description: Create a new scribe session for a clinical encounter
      operationId: createSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Sessions
      summary: List sessions
      description: Retrieve user's scribe sessions
      operationId: listSessions
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [initialized, active, completed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  total:
                    type: integer
...
  /api/sessions/{sessionId}/patient:
    post:
      tags:
        - Sessions
      summary: Bind patient to session
      description: Associate a patient with a scribe session
      operationId: bindPatient
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindPatientRequest'
      responses:
        '200':
          description: Patient bound to session
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session or patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/{sessionId}/note-sections/{sectionKey}:
    patch:
      tags:
        - Sessions
      summary: Update note section
      description: Update a specific note section in a session
      operationId: updateNoteSection
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sectionKey
          in: path
          required: true
          schema:
            type: string
            description: Section key (e.g., 'chief_complaint', 'assessment')
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteSectionRequest'
      responses:
        '200':
          description: Section updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # RECORDING ENDPOINTS
  # ============================================================================
...
security:
  - BearerAuth: []